Make LZMA dictionary size customizable.
The XZ library was using a 64MB dictionary by default, which can easily
exceed the Java VM size on older devices. This gives consumers of the
library the flexibility to specify a custom dictionary size, with the
tradeoff that it will be possible that larger bits of content (i.e.
anything larger than the dictionary size) will not be retrievable from
the Zim file.